Move update channel setting to Cody & Sourcegraph
