+++ src/main/kotlin/com/sourcegraph/cody/config/SettingsModel.kt
@@ -1,6 +1,5 @@
 package com.sourcegraph.cody.config
 
-import com.sourcegraph.cody.config.ui.UpdateChannel
 import java.awt.Color
 
 data class SettingsModel(
@@ -14,5 +13,4 @@ data class SettingsModel(
     var customAutocompleteColor: Color? = null,
     var blacklistedLanguageIds: List<String> = listOf(),
     var shouldAcceptNonTrustedCertificatesAutomatically: Boolean = false,
-    var channel: UpdateChannel = UpdateChannel.Stable
 )
